create embedding folder.
