+++ tibra/embedding/aabb_tree.cpp
@@ -1,0 +1,56 @@
+// Author: Manuel Meßmer
+// Email: manuel.messmer@tum.de
+
+// Project includes
+#include "embedding/aabb_tree.h"
+
+bool AABB_tree::IsWithinBoundingBox(const std::array<double, 3>& rPoint){
+    if(   rPoint[0] < mLowerBound[0]
+        || rPoint[0] > mUpperBound[0]
+        || rPoint[1] < mLowerBound[1]
+        || rPoint[1] > mUpperBound[1]
+        || rPoint[2] < mLowerBound[2]
+        || rPoint[2] > mUpperBound[2])
+    {
+        return false;
+    }
+
+    return true;
+}
+
+
+std::vector<AABB_tree::IndexType> AABB_tree::Query(AABB_primitive_base& rAABB_primitive)
+{
+    std::vector<IndexType> stack;
+    stack.reserve(256);
+    stack.push_back(BaseTreeType::Root());
+
+    std::vector<IndexType> particles;
+
+    while (stack.size() > 0)
+    {
+        IndexType node = stack.back();
+        stack.pop_back();
+
+        // Copy the AABB_base.
+        auto& aabb = static_cast<AABB_primitive&>(BaseTreeType::Nodes()[node].aabb_base);
+
+        if (node == NULL_NODE) continue;
+
+        // Test for overlap between the AABBs.
+        if (rAABB_primitive.intersect(aabb) )
+        {
+            // Check that we're at a leaf node.
+            if (BaseTreeType::Nodes()[node].isLeaf())
+            {
+                particles.push_back(BaseTreeType::Nodes()[node].particle);
+            }
+            else
+            {
+                stack.push_back(BaseTreeType::Nodes()[node].left);
+                stack.push_back(BaseTreeType::Nodes()[node].right);
+            }
+        }
+    }
+    return particles;
+}
