+++ tibra/tests/cpp_tests/test_point_classifier.cpp
@@ -8,7 +8,7 @@
 
 #include "geometries/triangle_mesh.h"
 #include "io/io_utilities.h"
-#include "utilities/geometrical_entity_classifier.h"
+#include "embedding/geometrical_entity_classifier.h"
 #include "geometries/element_container.h"
 
 namespace Testing{
